Rename dotcomponents packages to dotcomrendering
Also make facia package structure the same as article with regards to dotcomrendering services like `FaciaPicker`.
