+++ facia/app/services/dotcomrendering/FaciaPicker.scala
@@ -1,4 +1,4 @@
-package utils
+package services.dotcomrendering
 
 import common.GuLogging
 import experiments.{ActiveExperiments, DCRFronts}
