+++ facia/app/services/dotcomrendering/RenderType.scala
@@ -1,4 +1,4 @@
-package utils
+package services.dotcomrendering
 
 sealed trait RenderType
 case object RemoteRender extends RenderType
